projectService/requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
mysqlclient==2.1.1
protobuf==4.22.1
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==1.3.20
